Move and fix get_tile_bounds
